moved components out of the pages folder
