+++ src/components/icons/check.tsx
